+++ kiltergpt/data/tokenizer.py
@@ -193,12 +193,23 @@ class Tokenizer:
             t = self.pad(t, pad)
         return t
 
-    def onehot(self, frames: str) -> torch.Tensor:
+    def onehot(self, frames: str | torch.Tensor) -> torch.Tensor:
         """Save presence/absence of each hold in a one-hot tensor"""
+        if isinstance(frames, str):
+            return self._onehot_from_string(frames)
+        return self._onehot_from_tensor(frames)
+
+    def _onehot_from_string(self, frames: str) -> torch.Tensor:
         t = torch.zeros(len(self.encode_map), dtype=torch.long)
         for token in self.split_tokens(frames):
             if token.startswith("p"):
                 t[self.encode_map[token]] = 1
+        return t
+
+    def _onehot_from_tensor(self, encoded_frames: torch.Tensor) -> torch.Tensor:
+        t = torch.zeros(len(self.encode_map), dtype=torch.long)
+        encoded_frames = encoded_frames[torch.isin(encoded_frames, self.hold_token_ids)]
+        t[encoded_frames] = 1
         return t
 
     def decode(self, x: torch.Tensor, clean: bool = False) -> list | tuple: